refactor: code smell 제거
