+++ src/main/resources/static/docs/coupon.html
@@ -461,7 +461,7 @@ body.book #toc,body.book #preamble,body.book h1.sect0,body.book .sect1>h2{page-b
 <div class="content">
 <pre class="highlight nowrap"><code class="language-http" data-lang="http">POST /admins/coupons HTTP/1.1
 Content-Type: application/json;charset=UTF-8
-Content-Length: 192
+Content-Length: 200
 Host: localhost:8080
 
 {
@@ -483,7 +483,7 @@ Vary: Origin
 Vary: Access-Control-Request-Method
 Vary: Access-Control-Request-Headers
 Content-Type: application/json
-Content-Length: 62
+Content-Length: 64
 
 {
   "message" : "쿠폰의 이름이 중복되었습니다."
@@ -514,7 +514,7 @@ Vary: Origin
 Vary: Access-Control-Request-Method
 Vary: Access-Control-Request-Headers
 Content-Type: application/json
-Content-Length: 56
+Content-Length: 58
 
 {
   "message" : "존재하지 않는 쿠폰입니다."
@@ -546,7 +546,7 @@ Vary: Origin
 Vary: Access-Control-Request-Method
 Vary: Access-Control-Request-Headers
 Content-Type: application/json
-Content-Length: 56
+Content-Length: 58
 
 {
   "message" : "존재하지 않는 쿠폰입니다."
@@ -569,7 +569,7 @@ Content-Length: 56
 <div class="content">
 <pre class="highlight nowrap"><code class="language-http" data-lang="http">POST /coupons/search HTTP/1.1
 Content-Type: application/json;charset=UTF-8
-Content-Length: 84
+Content-Length: 88
 Host: localhost:8080
 
 {
@@ -627,7 +627,7 @@ Content-Length: 3
 <div id="footer">
 <div id="footer-text">
 Version 0.0.1-SNAPSHOT<br>
-Last updated 2023-11-13 16:03:53 +0900
+Last updated 2023-11-15 12:06:59 +0900
 </div>
 </div>
 </body>
