+++ src/main/resources/static/docs/notification.html
@@ -473,7 +473,7 @@ Vary: Origin
 Vary: Access-Control-Request-Method
 Vary: Access-Control-Request-Headers
 Content-Type: application/json
-Content-Length: 64
+Content-Length: 66
 
 {
   "message" : "이미 콕 알림을 보낸 대상입니다."
@@ -487,7 +487,7 @@ Content-Length: 64
 <div id="footer">
 <div id="footer-text">
 Version 0.0.1-SNAPSHOT<br>
-Last updated 2023-11-10 21:12:31 +0900
+Last updated 2023-11-15 12:06:59 +0900
 </div>
 </div>
 </body>
